# set `COVERAGE_CORE=sysmon` to improve tests' performance in 3.12